creator/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==1.10.14
python-dotenv==1.0.1
pytest==8.3.2
